+++ R/data.R
@@ -1,5 +1,5 @@
 globalVariables(c("protein_sample_info", "protein_data", "batch_indicator",
-    "signature_data", "bladder_meta", "bladder_data"))
+    "signature_data", "bladder_meta", "bladder_data", "merged_IDs"))
 
 #' Batch and Condition indicator for protein expression data
 #'
@@ -94,6 +94,25 @@ bladder_data_upload <- function() {
     return(se_object)
 }
 
+#' BMI and matched sample names for TB data
+#'
+#' This is support data for the TB data set that contains the BMI data and ID
+#' numbers from both the curatedTBData database and the original study the data
+#' was used in
+#'
+#' @name merged_IDs
+#' @docType data
+#' @format A data frame with 91 rows and 3 columns
+#' \describe{
+#'     \item{subjectID_curatedTBData}{Subject ID found in curatedTBData}
+#'     \item{subjectID_TB_Paper}{Subject ID in the original paper}
+#'     \item{BMI}{subject's BMI from the original study}
+#' }
+#' @keywords datasets
+#' @usage data(merged_IDs)
+"merged_IDs"
+
+
 #' TB data upload
 #' This function uploads the TB data set from the curatedTBData package.
 #'
@@ -169,38 +188,12 @@ tb_data_upload <- function() {
 #' @return dataframe provided as input with BMI info added
 #'
 BMI_data <- function(meta) {
-    subjectID_curatedTBData <- c("GSM4609427", "GSM4609429", "GSM4609424",
-        "GSM4609396", "GSM4609417", "GSM4609387", "GSM4609415", "GSM4609408",
-        "GSM4609426", "GSM4609412", "GSM4609400", "GSM4609423", "GSM4609416",
-        "GSM4609425", "GSM4609420", "GSM4609386", "GSM4609422", "GSM4609406",
-        "GSM4609419", "GSM4609414", "GSM4609389", "GSM4609432", "GSM4609428",
-        "GSM4609397", "GSM4609421", "GSM4609410", "GSM4609413", "GSM4609411",
-        "GSM4609433", "GSM4609431", "GSM4609430", "GSM4609418", "GSM4609407",
-        "GSM4609402", "GSM4609404", "GSM4609398", "GSM4609394", "GSM4609392",
-        "GSM4609399", "GSM4609391", "GSM4609388", "GSM4609409", "GSM4609403",
-        "GSM4609395", "GSM4609393", "GSM4609390", "GSM4609405", "GSM4609401",
-        "GSM2712704", "GSM2712705", "GSM2712706", "GSM2712677", "GSM2712707",
-        "GSM2712696", "GSM2712698", "GSM2712699", "GSM2712708", "GSM2712700",
-        "GSM2712709", "GSM2712710", "GSM2712711", "GSM2712678", "GSM2712713",
-        "GSM2712714", "GSM2712679", "GSM2712715", "GSM2712680", "GSM2712681",
-        "GSM2712702", "GSM2712716", "GSM2712717", "GSM2712682", "GSM2712718",
-        "GSM2712683", "GSM2712684", "GSM2712692", "GSM2712685", "GSM2712686",
-        "GSM2712719", "GSM2712693", "GSM2712703", "GSM2712687", "GSM2712694",
-        "GSM2712697", "GSM2712688", "GSM2712689", "GSM2712690", "GSM2712695",
-        "GSM2712691", "GSM2712676", "GSM2712701")
-    BMI <- c(13.00, 13.61, 13.72, 13.87, 13.88, 13.92, 13.95, 13.95, 14.04,
-        14.11, 14.21, 14.38, 14.43, 14.77, 14.80, 14.81, 15.01, 15.02, 15.02,
-        15.02, 15.09, 15.17, 15.59, 15.61, 15.66, 15.68, 15.72, 15.75, 15.81,
-        15.81, 15.82, 15.86, 19.09, 19.52, 21.01, 21.26, 22.31, 22.44, 22.81,
-        22.87, 24.15, 26.31, 27.20, 27.44, 27.81, 28.12, 30.83, 33.16, 28.80,
-        24.80, 24.30, 21.50, 26.30, 15.10, 14.30, 15.60, 23.50, 15.10, 24.10,
-        38.70, 23.80, 19.40, 26.00, 21.80, 21.70, 22.30, 19.20, 24.90, 14.20,
-        22.00, 25.30, 20.10, 25.00, 20.40, 19.20, 14.30, 18.90, 20.30, 21.70,
-        17.90, 15.60, 18.60, 16.50, 11.50, 23.20, 26.00, 19.30, 21.40, 19.00,
-        20.90, 13.80)
+    data("merged_IDs", envir = environment())
+    subjectID_curatedTBData <- merged_IDs$subjectID_curatedTBData
+    BMI <- merged_IDs$BMI
     IDs <- data.frame(subjectID_curatedTBData, BMI)
     meta$BMI <- rep(NA, length(meta$Experiment))
-    for (i in seq_len(rownames(meta))){
+    for (i in seq_len(nrow(meta))){
         position <- which(IDs$subjectID_curatedTBData == rownames(meta)[i])
         meta$BMI[i] <- IDs$BMI[position]
     }